requirements.txt
customtkinter==5.2.2 
scrapetube==2.5.1
openai==1.48.0 
pillow==10.4.0
youtube_transcript_api==0.6.2
tiktoken==0.8.0